backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
stripe>=12.0.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
aiosqlite>=0.21.0
pybaseball>=2.2.7
beautifulsoup4>=4.13.0
lxml>=5.3.0
cachetools>=5.5.0
tenacity>=9.0.0
python-multipart>=0.0.20
slowapi>=0.1.9
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx[http2]>=0.28.0
